+++ crypto/blockcipher/blockcipher.d
@@ -5,5 +5,5 @@ public interface BlockCipher
     public void encrypt(ubyte[] message);
     public void decrypt(ubyte[] cipher);
 
-    @property public const uint blockSize();
+    @property public const size_t blockSize();
 }
